Refactor Database struct into SystemCatalog and QueryContext
